+++ vtm-android-example/src/org/oscim/android/test/S3DBMapActivity.java
@@ -19,7 +19,6 @@ package org.oscim.android.test;
 import android.os.Bundle;
 
 import org.oscim.android.cache.TileCache;
-import org.oscim.layers.tile.TileLayer;
 import org.oscim.layers.tile.buildings.S3DBLayer;
 import org.oscim.layers.tile.vector.labeling.LabelLayer;
 import org.oscim.theme.VtmThemes;
@@ -47,8 +46,8 @@ public class S3DBMapActivity extends BaseMapActivity {
             mS3dbCache.setCacheSize(512 * (1 << 10));
             ts.setCache(mS3dbCache);
         }
-        TileLayer tl = new S3DBLayer(mMap, ts, true, false);
-        mMap.layers().add(tl);
+
+        mMap.layers().add(new S3DBLayer(mMap, ts));
         mMap.layers().add(new LabelLayer(mMap, mBaseLayer));
     }
