+++ vtm-themes/resources/assets/vtm/default.xml
@@ -560,13 +560,25 @@
           </m> -->
 
         <!-- building -->
-        <m k="building">
-            <m zoom-min="14">
-                <m closed="yes">
-                    <area fade="14" use="building" />
+        <m k="building|building:part">
+            <m k="building">
+                <m zoom-min="14">
+                    <m closed="yes">
+                        <area fade="14" use="building" />
+                    </m>
+                    <m closed="no">
+                        <line fade="14" use="building" />
+                    </m>
                 </m>
-                <m closed="no">
-                    <line fade="14" use="building" />
+            </m>
+            <m k="building:part">
+                <m zoom-min="16">
+                    <m closed="yes">
+                        <area fade="16" use="building" />
+                    </m>
+                    <m closed="no">
+                        <line fade="16" use="building" />
+                    </m>
                 </m>
             </m>
             <m zoom-min="17">
