+++ vtm-web/src/com/badlogic/gdx/backends/Gdx.gwt.xml
@@ -3,6 +3,7 @@
     "http://google-web-toolkit.googlecode.com/svn/tags/2.5.1/distro-source/core/src/gwt-module.dtd">
 <module rename-to='com.badlogic.gdx.backends.gwt'>
     <inherits name='com.google.gwt.user.User' />
+    <inherits name='com.google.gwt.i18n.I18N' />
     <!-- Inherit edited chrome theme ("gwt"-prefixed classes only) for a little bit of default styling in the text input dialogs -->
     <inherits name='com.badlogic.gdx.backends.gwt.theme.chrome.Chrome' />
     <inherits name="com.google.gwt.http.HTTP" />
