+++ vtm/src/org/oscim/layers/tile/buildings/S3DBTileLoader.java
@@ -1,3 +1,20 @@
+/*
+ * Copyright 2014-2015 Hannes Janetzek
+ * Copyright 2016-2017 devemux86
+ *
+ * This file is part of the OpenScienceMap project (http://www.opensciencemap.org).
+ *
+ * This program is free software: you can redistribute it and/or modify it under the
+ * terms of the GNU Lesser General Public License as published by the Free Software
+ * Foundation, either version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY
+ * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
+ * PARTICULAR PURPOSE. See the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License along with
+ * this program. If not, see <http://www.gnu.org/licenses/>.
+ */
 package org.oscim.layers.tile.buildings;
 
 import org.oscim.backend.canvas.Color;
@@ -15,8 +32,6 @@ import org.oscim.tiling.QueryResult;
 import org.oscim.tiling.TileSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import static org.oscim.layers.tile.buildings.S3DBLayer.getMaterialColor;
 
 class S3DBTileLoader extends TileLoader {
     static final Logger log = LoggerFactory.getLogger(S3DBTileLoader.class);
@@ -115,11 +130,11 @@ class S3DBTileLoader extends TileLoader {
 
         int c = 0;
         if (element.tags.containsKey(COLOR_KEY)) {
-            c = S3DBLayer.getColor(element.tags.getValue(COLOR_KEY), isRoof);
+            c = S3DBRenderer.getColor(element.tags.getValue(COLOR_KEY), isRoof);
         }
 
         if (c == 0 && element.tags.containsKey(MATERIAL_KEY)) {
-            c = getMaterialColor(element.tags.getValue(MATERIAL_KEY), isRoof);
+            c = S3DBRenderer.getMaterialColor(element.tags.getValue(MATERIAL_KEY), isRoof);
         }
 
         if (c == 0) {
